add minigun from sc
implemented as WeaponCustom parameters.

differences from the sven version:
- fires about twice as fast
- you can jump a little bit and move slightly faster
- it doesn't have its own stash of 556 ammo

+++ cl_dll/com_weapons.cpp
@@ -237,7 +237,8 @@ UTIL_SharedRandomFloat
 float UTIL_SharedRandomFloat( unsigned int seed, float low, float high )
 {
 	//
-	unsigned int range;
+	//unsigned int range;
+	float range;
 
 	U_Srand( (int)seed + *(int *)&low + *(int *)&high );
